Try-catch on channel.js

+++ Utils/channel.js
@@ -1,13 +1,22 @@
 async function deleteChannel(channel) {
+    if (!channel) {
+        console.error("Channel is undefined.");
+        return;
+    }
+
     channel.delete()
         .then(() => console.log(`Successfully deleted channel ${channel.name}`))
-        .catch(console.error);
+        .catch(error => console.error(`Failed to delete channel ${channel.name}: ${error}`));
 }
 
+
 async function disconnect(user) {
-    user.voice.disconnect()
-        .then(() => console.log(`Successfully disconnected user ${user.displayName}`))
-        .catch(console.error);
+    try {
+        await user.voice.disconnect();
+        console.log(`Successfully disconnected user ${user.displayName}`);
+    } catch (error) {
+        console.error(`Failed to disconnect user ${user.displayName}: ${error}`);
+    }
 }
 
 async function move(user, channel) {
